refactor: 통계기능 추가를 위해 app.js분리
